Classes seperated and Ui was made more dynamic
